refactor(CantonCallBot): now is encapsulated
We can also start it with different scripts flask_app and main. It was
important to separate this entrypoints and the core logic/helpers
