+++ CantonCallBot/requirements.txt
@@ -1,0 +1,141 @@
+aiohttp==3.9.3
+aiosignal==1.3.1
+annotated-types==0.6.0
+anyio==4.3.0
+asgiref==3.8.1
+attrs==23.2.0
+backoff==2.2.1
+bcrypt==4.1.2
+beautifulsoup4==4.12.3
+bidict==0.23.1
+blinker==1.7.0
+bs4==0.0.2
+build==1.1.1
+cachetools==5.3.3
+certifi==2024.2.2
+cffi==1.16.0
+charset-normalizer==3.3.2
+chroma-hnswlib==0.7.3
+chromadb==0.4.24
+click==8.1.7
+cohere==5.1.1
+coloredlogs==15.0.1
+colorlog==6.8.2
+dataclasses-json==0.6.4
+Deprecated==1.2.14
+distro==1.9.0
+dnspython==2.6.1
+eventlet==0.35.2
+fastapi==0.110.0
+filelock==3.13.1
+Flask==3.0.2
+Flask-Cors==4.0.0
+Flask-SocketIO==5.3.6
+flatbuffers==24.3.7
+frozenlist==1.4.1
+fsspec==2024.3.1
+google-auth==2.29.0
+googleapis-common-protos==1.63.0
+greenlet==3.0.3
+groq==0.4.2
+grpcio==1.62.1
+h11==0.14.0
+httpcore==1.0.4
+httptools==0.6.1
+httpx==0.27.0
+huggingface-hub==0.21.4
+humanfriendly==10.0
+idna==3.6
+importlib-metadata==6.11.0
+importlib_resources==6.4.0
+itsdangerous==2.1.2
+Jinja2==3.1.3
+jsonpatch==1.33
+jsonpointer==2.4
+kubernetes==29.0.0
+langchain==0.1.13
+langchain-community==0.0.29
+langchain-core==0.1.33
+langchain-groq==0.0.1
+langchain-text-splitters==0.0.1
+langchainhub==0.1.15
+langsmith==0.1.31
+MarkupSafe==2.1.5
+marshmallow==3.21.1
+mmh3==4.1.0
+monotonic==1.6
+mpmath==1.3.0
+multidict==6.0.5
+mypy-extensions==1.0.0
+networkx==3.2.1
+numpy==1.26.4
+oauthlib==3.2.2
+onnxruntime==1.17.1
+openai==1.14.2
+opentelemetry-api==1.23.0
+opentelemetry-exporter-otlp-proto-common==1.23.0
+opentelemetry-exporter-otlp-proto-grpc==1.23.0
+opentelemetry-instrumentation==0.44b0
+opentelemetry-instrumentation-asgi==0.44b0
+opentelemetry-instrumentation-fastapi==0.44b0
+opentelemetry-proto==1.23.0
+opentelemetry-sdk==1.23.0
+opentelemetry-semantic-conventions==0.44b0
+opentelemetry-util-http==0.44b0
+orjson==3.9.15
+overrides==7.7.0
+packaging==23.2
+posthog==3.5.0
+protobuf==4.25.3
+pulsar-client==3.4.0
+pyasn1==0.5.1
+pyasn1-modules==0.3.0
+pycparser==2.21
+pydantic==2.6.4
+pydantic_core==2.16.3
+pydub==0.25.1
+pygame==2.5.2
+PyPika==0.48.9
+pyproject_hooks==1.0.0
+python-dateutil==2.9.0.post0
+python-dotenv==1.0.1
+python-engineio==4.9.0
+python-socketio==5.11.1
+PyYAML==6.0.1
+regex==2023.12.25
+requests==2.31.0
+requests-oauthlib==2.0.0
+rsa==4.9
+safetensors==0.4.2
+scipy==1.12.0
+simple-websocket==1.0.0
+six==1.16.0
+sniffio==1.3.1
+sounddevice==0.4.6
+soundfile==0.12.1
+soupsieve==2.5
+SpeechRecognition==3.10.1
+SQLAlchemy==2.0.28
+starlette==0.36.3
+sympy==1.12
+tenacity==8.2.3
+tokenizers==0.15.2
+torch==2.2.1
+torchaudio==2.2.1
+tqdm==4.66.2
+transformers==4.39.0
+typer==0.10.0
+types-requests==2.31.0.20240311
+typing-inspect==0.9.0
+typing_extensions==4.10.0
+urllib3==2.2.1
+uvicorn==0.29.0
+uvloop==0.19.0
+watchfiles==0.21.0
+websocket-client==1.7.0
+websockets==12.0
+Werkzeug==3.0.1
+wrapt==1.16.0
+wsproto==1.2.0
+yarl==1.9.4
+zipp==3.18.1
