+++ CantonCallBot/src/core/voice_output_elevenLabs.py
@@ -3,8 +3,10 @@ import scipy.io.wavfile
 import numpy as np
 from transformers import VitsTokenizer, VitsModel, set_seed
 import torch
+import os
 class TextToVoice:
-    voice_output_file_path: str = "resources/voice_outputs/synthesized_speech.wav"
+    print("Current Working Directory:", os.getcwd())
+    voice_output_file_path: str = "src/resources/voice_outputs/synthesized_speech.wav"
     def __init__(self, use_elevenlabs_api=False):
         self.use_elevenlabs_api = use_elevenlabs_api
         if not use_elevenlabs_api:
@@ -27,6 +29,7 @@ class TextToVoice:
         waveform = outputs.waveform[0]
         waveform_np = waveform.cpu().numpy()
         waveform_int16 = np.int16(waveform_np * np.iinfo(np.int16).max)
+        print("Current Working Directory:", os.getcwd())
         scipy.io.wavfile.write(self.voice_output_file_path, rate=self.model.config.sampling_rate, data=waveform_int16)
 
     def _generate_with_elevenlabs(self, text: str):
@@ -49,6 +52,7 @@ class TextToVoice:
         if response.status_code == 200:
             # Assuming the API returns a WAV file directly, you'll need to adjust this
             # part if the API returns a link to the file or the file in a different format
+            print("Current Working Directory:", os.getcwd())
             with open(self.voice_output_file_path, "wb") as f:
                 f.write(response.content)
         else:
