+++ CantonCallBot/src/resources/voice_inputs/whole_rec/received_audio_whole_0_16.wav
